Moving storeImage to Image model
