+++ app/Http/Controllers/ContentController.php
@@ -133,7 +133,7 @@ class ContentController extends Controller
 
         if ($request->hasFile('file')) {
             
-            $filename = $this->storeImage($request->file('file'), $type);
+            $filename = Image::storeImage($request->file('file'));
             $content->images()->create(['filename' => $filename]);
 
         }
@@ -210,7 +210,7 @@ class ContentController extends Controller
         
         if ($request->hasFile('file')) {
             
-            $filename = $this->storeImage($request->file('file'), $type);
+            $filename = Image::storeImage($request->file('file'));
             $content->images()->update(['filename' => $filename]);
 
         }
@@ -245,47 +245,6 @@ class ContentController extends Controller
             ->route('content.show', [$type, $content])
             ->with('info', trans("content.update.info", ['title' => $content->title]));
 
-    }
-
-
-    public function storeImage($file, $type)
-    {
-
-        $fileName = $file->getClientOriginalName();
-        $fileName = preg_replace('/\s+/', '-', $fileName);
-        $path = config('imagepresets.original.path');
-        
-        $file->move($path, $fileName);
-
-        Imageconv::make($path . $fileName)
-            ->resize(
-                config('imagepresets.small.width'),
-                config('imagepresets.small.height'),
-                function ($constraint) {
-                    $constraint->aspectRatio();
-            })
-            ->save(config('imagepresets.small.path') . $fileName);
-
-        Imageconv::make($path . $fileName)
-            ->resize(
-                config('imagepresets.medium.width'),
-                config('imagepresets.medium.height'),
-                function ($constraint) {
-                    $constraint->aspectRatio();
-            })
-            ->save(config('imagepresets.medium.path') . $fileName);
-
-        Imageconv::make($path . $fileName)
-            ->resize(
-                config('imagepresets.large.width'),
-                config('imagepresets.large.height'),
-                function ($constraint) {
-                    $constraint->aspectRatio();
-            })
-            ->save(config('imagepresets.large.path') . $fileName);
-
-        return $fileName;
-    
     }
 
     public function status($type, $id, $status)
